# docs/api.yaml
openapi: 3.0.3
info:
  version: "1.0.0"
  title: "Resource API"
  description: "API for managing resources such as user, visitor, appointment, car, detail, image, and price."
servers:
  - url: "http://localhost:8080/api"

    # TODO: Different Schema for Dto Request & Response    
paths:
  /admins:
    get:
      summary: "Get all admins"
      operationId: "getAllAdmins"
      responses:
        '200':
          description: "A list of admins"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'

    put:
      summary: "Save or update an admin"
      operationId: "saveOrUpdateAdmin"
      requestBody:
        description: "Admin to save or update"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Admin'
      responses:
        '200':
          description: "Admin saved or updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
  /admins/{id}:
    get:
      summary: "Get an admin by ID"
      operationId: "getAdminById"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: "A single admin"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: "Admin not found"
    delete:
      summary: "Delete an admin by ID"
      operationId: "deleteAdminById"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: "Admin deleted"
        '404':
          description: "Admin not found"
...
  /cars:
    get:
      summary: "Get all cars"
      operationId: "getAllCars"
      responses:
        '200':
          description: "A list of cars"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'
    put:
      summary: "Save or update a car"
      operationId: "saveOrUpdateCar"
      requestBody:
        description: "Car to save or update"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
      responses:
        '200':
          description: "Car saved or updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
  /cars/{id}:
    get:
      summary: "Get a car by ID"
      operationId: "getCarById"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: "A single car"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: "Car not found"
    delete:
      summary: "Delete a car by ID"
      operationId: "deleteCarById"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: "Car deleted"
        '404':
          description: "Car not found"
...
components:
  schemas:
    Admin:
      type: object
      properties:
        id:
          type: integer
          example: 1
        firstname:
          type: string
          example: "john"
        email:
          type: string
          example: "johndoe@example.com"
        tel:
          type: string
          example: "0340000000"
        password:
          type: string
          example: "password"
      example:
        $ref: '#/components/examples/AdminExample'
...
    Car:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Tesla"
        description:
          type: string
          example: "Tesla Model S"
        color:
          type: string
          example: "red"
        motor-type:
          type: integer
          example: "electric"
        power:
          type: string
          example: "250horsepower"
        status:
          type: string
          enum:
            - "pinned"
            - "not_pinned"
          example: "pinned"
        type:
          type: string
          example: "Electric sedan"
        detailsId:
          type: integer
          example: 1
        priceId:
          type: integer
          example: 1
      example:
        $ref: '#/components/examples/CarExample'
...
  examples:
    AdminExample:
      value:
        id: 1
        firstname: "john"
        email: "johndoe@example.com"
        tel: "0340000000"
        password: "********"
...
    CarExample:
      value:
        id: 1
        name: "Tesla"
        description: "Tesla Model S"
        color: "red"
        motor-type: "electric"
        power: "250horsepower"
        status: "pinned"
        type: "Electric sedan"
        detailsId: 1
        priceId: 1
...